Address code review feedback:
* Improve documentation of constrained_delaunay
* Remove constrained_delaunay from ops.py
* Move test to test_constructive
